Redondear numeros Abel Sabater

+++ src/main/java/org/example/calculadora/RedondeodeNumeros.java
@@ -11,7 +11,7 @@ public class RedondeodeNumeros {
         return num;
     }
 
-    public void setNum(int num) {
+    public void setNum(double num) {
         this.num = num;
     }
     public int numerordondejat(){
